move page header of repository page into a component
