+++ app/components/repositories/page_header_component.html.erb
@@ -1,0 +1,35 @@
+<%=
+  render Primer::OpenProject::PageHeader.new do |header|
+    header.with_title { t(
+      "repositories.named_repository",
+      vendor_name: @repository.class.vendor_name
+    ) }
+    header.with_breadcrumbs(breadcrumb_items)
+    if !@empty && User.current.allowed_in_project?(:browse_repository, @project)
+      header.with_action_icon_button(
+        tag: :a,
+        icon: :graph,
+        label: t(:label_statistics),
+        mobile_icon: :graph,
+        mobile_label: t(:label_statistics),
+        size: :medium,
+        href: stats_project_repository_path(@project),
+        aria: { label: t(:label_statistics) },
+        title: t(:label_statistics)
+      )
+    end
+    if User.current.allowed_in_project?(:manage_repository, @project)
+      header.with_action_icon_button(
+        tag: :a,
+        icon: :gear,
+        label: t(:label_setting_plural),
+        mobile_icon: :gear,
+        mobile_label: t(:label_setting_plural),
+        size: :medium,
+        href: project_settings_repository_path(@project),
+        aria: { label: t(:label_setting_plural) },
+        title: t(:label_setting_plural)
+      )
+    end
+  end
+%>
