+++ frontend/src/global_styles/openproject/_scm.sass
@@ -73,17 +73,6 @@ li.change
 .repository--revision-toolbar
   margin-top: 3rem
 
-.repository-breadcrumbs
-  font-size: 0.9rem
-  margin: 15px 0
-
-.repository-breadcrumbs--home
-  font-size: 0.75rem
-
-.repository-bradcrumbs--identifier
-  display: inline-block
-  margin-left: 10px
-
 .repository-checkout-container
   width: 50%
   display: flex
